requirements.txt
pandas
requests
sqlalchemy
mysql-connector-python